requirements.txt
langgraph>=0.3.26
langgraph-supervisor>=0.0.29
langchain>=0.2.16
langchain-openai
pydantic>=2.8.2
requests>=2.32.3
openai
python-dotenv
tiktoken
